+++ CHANGES.md
@@ -1,11 +1,21 @@
 CHANGES
 ========
 
-## 2.4.4-RC
+## 2.4.5-RC
 
+* 
+
+## 2.4.4
+
+* Feature: Autoformat Markdown Table
+* Feature: highlight.js Theme Selector
+* Fix: The bug of updating numbering list by codemirror
 * Fix: Template LangProcessor doesn't work
     * Introduced by 2.4.0
 * Support: Apply ESLint
+* Support: Upgrade libs
+    * react, react-dom
+    * codemirror, react-codemirror2
 
 ## 2.4.3